# docs/swagger.yaml
basePath: /api
definitions:
  crud.Access-string:
    properties:
      actId:
        type: string
      resId:
        type: string
      usrId:
        type: string
    required:
    - actId
    - resId
    - usrId
    type: object
  crud.AccessList-string:
    properties:
      policies:
        items:
          $ref: '#/definitions/crud.Access-string'
        type: array
    type: object
  crud.AccessListUpdate:
    properties:
      policies:
        items:
          $ref: '#/definitions/crud.AccessUpdate'
        type: array
    type: object
  crud.AccessUpdate:
    properties:
      new:
        $ref: '#/definitions/crud.Access-string'
      old:
        $ref: '#/definitions/crud.Access-string'
    required:
    - new
    - old
    type: object
info:
  contact:
    email: edgardanielgd123@gmail.com
    name: API Support
  description: UIdealist Authorization project API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
  /v1/check:
    post:
      consumes:
      - application/json
      description: Checks against a RBAC policy if a user is allowed to access a resource.
      parameters:
      - description: Policies to check
        in: body
        name: policies
        required: true
        schema:
          $ref: '#/definitions/crud.AccessList-string'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Check if resource is accessible by user
      tags:
      - Access
  /v1/grant:
    post:
      consumes:
      - application/json
      description: Alters a RBAC policy to grant access to a resource.
      parameters:
      - description: Policies to grant
        in: body
        name: policies
        required: true
        schema:
          $ref: '#/definitions/crud.AccessList-string'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Grant access to resource
      tags:
      - Access
  /v1/revoke:
    post:
      consumes:
      - application/json
      description: Alters a RBAC policy to revoke access to a resource.
      parameters:
      - description: Policies to revoke
        in: body
        name: policies
        required: true
        schema:
          $ref: '#/definitions/crud.AccessList-string'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Revoke access to resource
      tags:
      - Access
  /v1/update:
    post:
      consumes:
      - application/json
      description: Alters a RBAC policy to update access to a resource.
      parameters:
      - description: Policies to update
        in: body
        name: policies
        required: true
        schema:
          $ref: '#/definitions/crud.AccessListUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Update access to resource
      tags:
      - Access
swagger: "2.0"